+++ .vscode/settings.json
@@ -19,4 +19,24 @@
   },
   "editor.insertSpaces": true,
   "editor.tabSize": 2,
+  "workbench.colorCustomizations": {
+    "activityBar.activeBackground": "#e0943d",
+    "activityBar.background": "#e0943d",
+    "activityBar.foreground": "#15202b",
+    "activityBar.inactiveForeground": "#15202b99",
+    "activityBarBadge.background": "#c8f6e1",
+    "activityBarBadge.foreground": "#15202b",
+    "commandCenter.border": "#15202b99",
+    "sash.hoverBorder": "#e0943d",
+    "statusBar.background": "#cb7b20",
+    "statusBar.foreground": "#15202b",
+    "statusBarItem.hoverBackground": "#9f6019",
+    "statusBarItem.remoteBackground": "#cb7b20",
+    "statusBarItem.remoteForeground": "#15202b",
+    "titleBar.activeBackground": "#cb7b20",
+    "titleBar.activeForeground": "#15202b",
+    "titleBar.inactiveBackground": "#cb7b2099",
+    "titleBar.inactiveForeground": "#15202b99"
+  },
+  "peacock.color": "#cb7b20",
 }
