+++ package.json
@@ -34,6 +34,7 @@
     "eslint": "^8.46.0",
     "nuxt": "^3.6.5",
     "nuxt-icon": "^0.4.2",
+    "nuxt-og-image": "^2.0.13",
     "prisma": "^5.0.0",
     "rehype-katex": "^6.0.3",
     "remark-math": "^5.1.1",
